refactor vote creation tests to more test classes
