+++ src/test/java/org/rulez/demokracia/pdengine/VoteCreationTest.java
@@ -3,7 +3,6 @@ package org.rulez.demokracia.pdengine;
 import static org.junit.Assert.*;
 
 import java.time.Instant;
-import java.util.Arrays;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -22,13 +21,6 @@ public class VoteCreationTest extends CreatedDefaultVoteRegistry{
 	@Before
 	public void setUp() throws ReportedException {
 		super.setUp();
-	}
-
-	private String createLongString(int length) {
-		char[] charArray = new char[length];
-        Arrays.fill(charArray, 'w');
-        String str256 = new String(charArray);
-		return str256;
 	}
 
 	@Test
@@ -104,244 +96,4 @@ public class VoteCreationTest extends CreatedDefaultVoteRegistry{
 		assertNotNull(adminInfo.voteId);
 		assertEquals(adminInfo.adminKey, voteManager.getVote(adminInfo.voteId).adminKey);
 	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void vote_name_can_contain_scpaces() throws ReportedException {
-		voteName = "This contains spaces";
-		VoteAdminInfo voteadm = createAVote();
-		assertEquals(voteName, voteManager.getVote(voteadm.voteId).name);
-		
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")	@Test
-	public void vote_name_cannot_be_null() throws ReportedException {
-		voteName = null;
-		assertThrows(
-			() -> createAVote()
-		).assertMessageIs("vote name is null");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void vote_name_cannot_contain_tabs() {
-		voteName = "thiscontainstab\t";
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("invalid characters in vote name");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void votename_max_length_is_255_characters() throws ReportedException {
-		int length = 255;
-		String str255 = createLongString(length);
- 		voteName = str255;
-
- 		createAVote();
- 		voteName = str255+"w";
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too long: vote name");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void minimum_vote_name_length_is_3() throws ReportedException {
-		voteName = "aaa";
-		createAVote();
-		voteName = "aa";
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too short: vote name");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void votename_can_contain_local_characters() throws ReportedException {
-		voteName = "ThisConatinsLocaCharséűáőúöüóíÉÁŰŐÚÖÜÓÍ";
-
-		VoteAdminInfo secondVote = createAVote();
-
-		assertEquals(voteName, voteManager.getVote(secondVote.voteId).name);
-
-	}
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void neededAssurances_is_checked_not_to_contain_strings_longer_than_255() throws ReportedException {
-        int length = 255;
-        String str255 = createLongString(length);
-
-        neededAssurances.add(str255);
-
-        createAVote();
-        String str256 = str255 + "w";
-        neededAssurances.add(str256);
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too long: needed assurance name");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void neededAssurances_is_checked_not_to_contain_strings_shorter_than_3() throws ReportedException {
-        neededAssurances.add("aaa");
-
-        createAVote();
-        neededAssurances.add("aa");
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too short: needed assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void needed_assurances_should_not_contain_space() {
-        neededAssurances.add("This contains space");
-
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("invalid characters in needed assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void needed_assurances_should_not_contain_tab() {
-        neededAssurances.add("thiscontainstab\t");
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("invalid characters in needed assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void needed_assurances_should_not_contain_empty_string() {
-        neededAssurances.add("");
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too short: needed assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void needed_assurances_should_not_be_null() {
-        neededAssurances.add(null);
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("needed assurance name is null");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void needed_assurances_can_contain_local_characters() throws ReportedException {
-        neededAssurances.add("ThisConatinsLocaCharséűáőúöüóíÉÁŰŐÚÖÜÓÍ");
-        
-        createAVote();
-        assertEquals(voteName, voteManager.getVote(adminInfo.voteId).name);
-
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-	public void countedAssurances_is_checked_not_to_contain_strings_longer_than_255() throws ReportedException {
-        int length = 255;
-        String str255 = createLongString(length);
-
-        countedAssurances.add(str255);
-
-        createAVote();
-        String str256 = str255 + "w";
-        countedAssurances.add(str256);
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too long: counted assurance name");
-	}
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void countedAssurances_is_checked_not_to_contain_strings_shorter_than_3() throws ReportedException {
-        countedAssurances.add("aaa");
-
-        createAVote();
-        countedAssurances.add("aa");
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("string too short: counted assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void counted_assurances_should_not_contain_space() {
-        countedAssurances.add("This contains space");
-
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("invalid characters in counted assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void counted_assurances_should_not_contain_tab() {
-        countedAssurances.add("thiscontainstab\t");
-		assertThrows(
-				() -> createAVote()
-			).assertMessageIs("invalid characters in counted assurance name");
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void counted_assurances_can_contain_local_characters() throws ReportedException {
-        countedAssurances.add("ThisConatinsLocaCharséűáőúöüóíÉÁŰŐÚÖÜÓÍ");
-        
-        createAVote();
-        assertEquals(voteName, voteManager.getVote(adminInfo.voteId).name);
-
-    }
-
-	@tested_feature("Manage votes")
-	@tested_operation("create vote")
-	@tested_behaviour("formally validates all inputs")
-	@Test
-    public void counted_assurance_can_be_empty() throws ReportedException {
-        countedAssurances.add("");
-        assertEquals(countedAssurances.size(),1);
-		VoteAdminInfo secondVote = createAVote();
-		assertTrue(voteManager.getVote(secondVote.voteId).countedAssurances.contains(null));
-    }
 }
